+++ server/utils/agents/abitat/plugins/cli.js
